added jumbotron to landing page. also got login button working
