+++ src/components/material-ui/Landing/Jumbotron.js
@@ -1,7 +1,20 @@
 import React from 'react'
 import Paper from 'material-ui/Paper'
-import RaisedButton from 'material-ui/RaisedButton'
-import FontIcon from 'material-ui/FontIcon'
+import LoginBtn from '../../../components/material-ui/Landing/LoginBtn'
+
+const Jumbotron = () => (
+  <Paper style={styles} key='Jumbotron'>
+    <section
+      id='intro-container'
+      style={intro}>
+    </section>
+    <div id='login-btn'>
+      <LoginBtn />
+    </div>
+  </Paper>
+)
+
+export default Jumbotron
 
 const styles = {
   height: 420,
@@ -19,40 +32,5 @@ const intro = {
   height: 220,
   width: 600,
   margin: '0 auto',
-  marginTop: 55
+  marginTop: 48
 }
-
-const btn = {
-  button: {
-    margin: 12,
-    backgroundColor: 'rgb(110, 206, 90)'
-  },
-  exampleImageInput: {
-    cursor: 'pointer',
-    position: 'absolute',
-    top: 0,
-    bottom: 0,
-    right: 0,
-    left: 0,
-    width: '100%',
-    opacity: 0
-  }
-}
-
-const Jumbotron = () => (
-  <Paper style={styles} key='Jumbotron'>
-    <section
-      id='intro-container'
-      style={intro}>
-    </section>
-    <div id='login-btn'>
-      <RaisedButton
-        label='Login with GitHub'
-        backgroundColor='#64DD17'
-        icon={<FontIcon className='muidocs-icon-custom-github' />}
-      />
-    </div>
-  </Paper>
-)
-
-export default Jumbotron
